Avoid re capture group limit of 100 by giving up on those examples.
So when more than 98 capture groups would be needed, it just classifies
all characters as anything ('.').

This is kind-of pathetic, but better than bombing out as it did.

Should improve as the algorithm improves.

Added new test for this behaviour and a few user agents (from Noam Ross ---
who found the problem: thanks, Noam) as an example file in rexpy/examples.