Added some special framework support for wildcard import for pytest.
It implements the referencepytest module by letting an instance of
the ReferenceTest class masquerade as a module. So that class needs to
provide some extra support to allow wildcard import, because it's not
really a module even though it's pretending to be one. This makes the
pretence a bit more complete.