Restructured the tdda repository so that it has a standard PyPi structure.
In particular, its 'root module' (tdda) is now a subdirectory (so
everything has moved down one level).

Also added a setup.py script so that it can package itself for PyPi.

Also modified the README files and __init__.py files to start to
make it so that it can document itself better (even for a binary
release).