More options for tdda command and minor clean-up.
The tdda command now accepts help and test and variants of those.

It is also tested by the tests.

Changed default (on the branch) for inc_rex to false and added options
to make this available ahead of merging into master.

Still need some tests for discover and verify with regular expressions.