Add new setting variableLengthGroups to rexpy's extract/Extractor.
This is False by default, retaining previous behaviour.

When set to True, rexpy will allow regular-expression fragments
to include optional bits on the end (which has not been the case
hitherto), e.g. 'https?'.

This is sometimes helpful, but not always. (In particular, it can
make Rexpy produce less general regular expressions when small
numbers of examples are provided). When it's a bit better, we will
probably add a command-line flag to control this, and perhaps make
it the default.

There are also minor enhancements in generating regular expressions,
so that ? is used instead of {0,1} (it wasn't previously actually
ever generated, but can be now), and also so that single repeats
of letters in single-character non-fixed fragments will be expanded
rather than using {2}, e.g. 'https?' will be used rather than 'ht{2}ps?'.