Update to support Jython; removed sqlite3 dependency.
To check the version of Python, there were checks for sys.version_info.major,
which works in CPython 2.7 and CPython 3.x, but not in Jython.
Changed to use sys.version_info[0].

It turns out that sqlite3, which was supposed to be an optional dependency,
was required for the tests to work; it has been made properly optional.

There was also an issue under Python3 triggering an warning in cases
in which pandas.isnull on empty iterables was returning an empty array,
and the truth value of this being tested. This is now handled differently
to avoid that (ambiguous) check.