Added support for running subsets of TDDA reference tests, rather than always running them all.
This is useful during development (of analysis and tests), and is also
useful for limited regeneration of results (when a change requires the
results to be regenerated, but only for a specific test or tests).

When run with the --tagged option on the command line, only tests (or
test classes) that have been decorated with @tag are run. The -1 option
is a shorthand for --tagged.

When run with the --istagged option on the command line, no tests are
run, but the names of all the classes that have been tagged are listed.

To tag a unittest test or class, use the @tag decorator, as in the
following example:

    from tdda.referencetest.referencetestcase import ReferenceTestCase, tag

    # tag an entire test class;
    # all tests in this class will run when --tagged is used.
    @tag
    class MyTests1(ReferenceTestCase):
        ...

    class MyTests2(ReferenceTestCase):
        # tag an individual test in a test class;
        # only test_a will run when --tagged is used.
        @tag
        def test_a(self):
            ...

        def test_b(self):
            ...

Tagged tests (for running subsets) is only available under unittest
at present, not under pytest yet.