When running verify detection, empty and remove the output file first.
This means that you get early notification if you specify an output file
that you can't write to. (It'll fail right at the start, rather than at
the end, after a lot of time-consuming processing).

It also means you don't get left with an out-of-date output file containing
detections from a previous run, if you run it and nothing is detected (and
therefore no output file is written).