Fixed bugs with epsilon in constraint verify, and with detection ok fields.
The epsilon flag was being ignored (both through the API and the command
line), because comparisons were always being done in 'closed' mode, and
there was no way to override that. Now, all comparisons are done in 'fuzzy'
mode (which means that they're 'closed' if epsilon is zero, and tolerant
otherwise). The default value for epsilon is now zero, so that, along with
the change above, means that the flag now works as expected.

The detection 'ok' fields were being created with the wrong type (should be
boolean, not real), and that's fixed now.

The 'nonnull' and 'nodups' ones were more fundamentally broken in that they
were being computed in a way that meant that sometimes they could be null
(NaN). That was wrong, and is now fixed too.