Added better control of rownumber in detection output, and boolean fields.
The rownumber produced by detection now takes account of both where the
data originated and where the detection output is being written to:

There are four cases

  1. CSV -> CSV.  The output field RowNumber is the row-number from the input.
  3. Feather -> CSV. The output field Index comes from the index on the input.
  2. CSV -> Feather. The output index comes from the thw-number from the input.
  4. Feather -> Feather. The output index comes from the index on the input.

Also added an optional --int flag (or boolean_ints=True via the API) for
detection, to write out boolean values as integers to CSV files, rather than
as "true" and "false" values. This is useful for cases where the CSV file
is going to be consumed by some other process that can't read booleans.