Fixed bug in tdda.referencetest.ReferenceTest.set_default_data_location.
It was supposed to apply at a per-class level (so if you call it on the
base ReferenceTest class, then it affects ALL tests in all subclasses, but
if you call it just on your subclass, then it only affects that subclass,
plus any sub-subclasses of that which don't themselves override it).

But that's now how it was implemented. It was always applying the defaults
to the base class, and using those in all subclasses. So that meant that
if you used on it a large project, with a number of different test-data
locations for different submodules, then you couldn't do that (except by
explcitly setting the data location within each individual instance, bacause
the class-based defaults weren't working right).

They're now working right. Also added an extra test to test this.