Modified the Pandas-based constraint commands to allow '-' as file.
A '-' means stdin on input, and stdout on output.

Signature for tdda discover is now:
    tdda discover input [output]
where both input and output can be '-'.

Signature for tdda verify is now:
    tdda verify input [constraints [output]]
where both input and output can be '-', and constraints is inferred from
input if it can be.

Signature for tdda detect is now
    tdda detect input [constraints [output]]
where both input and output can be '-'
where both input and output can be '-', and constraints is inferred from
input if it can be.

So these commands can now be used as follows:

    tdda discover - [-]            (discover from stdin, write json to stdout)
    tdda verify - constraints [-]  (read from stdin, verify against constraints)
    tdda detect - constraints [-]  (read from stdin, verify against constraints
                                    and write failures to stdout)

Also fixed issue with boolean fields that contain nulls. They were being
misinterpreted as being string fields, but then failing (because they
weren't really string fields). Booleans with missing values are now just
booleans. Tests for this now updated.