Check exit status, and require to be zero by default.
Allow NONZEROEXIT to specify that a non-zero exit status is OK on command line,
and add question to wizard for this.

Warn if any of the files specified are not available for copying, but
still generated the test script.

Remove existing reference files and script before generating, if they exist.