Fix various issues with gentest.
It now does uses substrings and patterns for exclusion patterns,
as appropriate, and (at least more often) gets the paths and
exclusions right for files with the same name.

It also now generates exclusion patterns for files, as well as for
STDOUT and STDERR. (It was generating them before, but failing to write
them out because of a failed lookup.)

There is also now support for a number of iterations to be passed in
and the wizard also offers this.