Reference test regeneration of results now takes account of stripping.
If a test is specified with lstrip or rstrip, then the appropriately-stripped
lines are written out to the regenerated result file if you run with the
--write-all or --write flags.

Previously it was writing them out raw (and its comparison was taking the
stripping into account, so everything 'worked', but it meant that reporting
on 'post-processed' differences was reporting more lines as being different
than it ought to have done - in tools like diff, opendiff, etc).

Also a small tweak to use OrderedDict more in the constraints module.
Makes no difference for python3, but makes API use of python2 more predictable.