For for potential hang in rexpy while computing coverage stats.
Fix in rexpy code for computing incremental coverage, to deal with some
edge cases that were hanging. It wasn't taking account of the possibility
that a *subset* of the expressions found could give 100% coverage, and
therefore some of the other expressions aren't contributing anything.

This couldn't happen in normal use, but could happen when called via API
by clients that are doing cleverer things about outliers.