Default to reset index on loading parquet file.
ddiff was behaving crazily because Snowflake
seems to create dataframes with odd indexes.

Now default to resetting in-place.

Should probably change dataframe comparison to do same.

Renamed the df attribute on Diffs objects to dfd,
since it's a DataFrameDiffs object.

Fixed a couple of doc strings Python 3.12 didn't like
(which might mess up the documentation) and also fixed
a regex it didn't like that was, in fact, incorrect.