Fixed test failures.
There is a problem if write_detection_reports is on and output_fields
are specified but do not include all the fields with failures.
Some of the failure detection code for report writing was being
called even when there were no reports to be written. This has
now been corrected. But a problem remains if reports are
turned on as above.

This is now logged as issue #41