Progress on new summary table for detect.
I think it's not more-or-less working, both for Pandas and databases.
Tests need to be updated and it all needs to be checked.

There is also an implementation of password_env_var as an alternative
to password in databse connection files (which needs to be tested manually
and by the addition of new automated tests)