Pandas metadata writing getting very close.
Four tests, each for 3 flavours of metadata written,
for four tiny datasets, with and without nulls,
and using nullable and non-nullable types.

Still need to decide which flavours to write by default
for pandas and whether to write the index column by default
(as Pandas itself does, but which really makes little sense).