fuzzy_nulls!
Extended pandas metadata tests further (still with extremely
simple cases) and added fuzzy_nulls as an option to
pandas dataframe comparisons (not yet to cli).

fuzzy_nulls defaults to False, but if set to object
will map nulls in object cols with strings to ''
and also the column they are being compared to,
even if this is string. If we to True, it does
the same for string columns.

This basically says "Pandas doesn't really understand
the difference between empty strings and nulls in CSV
files, so we;ll treat them as equivalent.